crystaledit: reduce compile warnings
--HG--
branch : stable

+++ Externals/crystaledit/editlib/tcl.cpp
@@ -109,7 +109,7 @@ ParseLineTcl (DWORD dwCookie, int nLineIndex, TEXTBLOCK * pBuf, int &nActualItem
   int nIdentBegin = -1;
   int nPrevI = -1;
   int I=0;
-  for (I = 0;; nPrevI = I, I = ::CharNext(pszChars+I) - pszChars)
+  for (I = 0;; nPrevI = I, I = static_cast<int>(::CharNext(pszChars+I) - pszChars))
     {
       if (I == nPrevI)
         {
